Remove internal from FlippableController to allow initialization in view model
Currently, device rotation restores the FlippableState to INITIALIZE. In order to persist state
across device rotation I would like to host the FlippableController in the view model so that
I can manager it there.

Alternatively, it would be nice to be able to implement my own remember where I'm able to provide a key
so that I can update the FlippableController when the key changes.